Added new setting:
JOHNNY_CACHE_DISABLE_SELECT

This is to disable selecting data from the cache but not to disable the
invalidation of the cache on update. This is useful for situations where
you have an "admin server" on it's own. You don't need/want to read from
the cache but you need to be able to invalidate it when data changes.